requirements.txt
amqp==5.2.0
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
boto3==1.28.79
botocore==1.31.79
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
convertdate==2.4.0
cron-descriptor==1.4.0
cryptography==41.0.5
cssselect2==0.7.0
cycler==0.12.1
Django==4.2.4
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-compat==1.0.15
django-js-asset==2.1.0
django-multiupload==0.6.1
django-picklefield==3.1
django-q==1.3.9
django-storages==1.14.2
django-summernote==0.8.20.0
django-timezone-field==6.1.0
django-tinymce==3.6.1
docopt==0.6.2
Flask==3.0.0
flower==2.0.1
fonttools==4.45.0
freetype-py==2.3.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
holidays==0.41
html2text==2020.1.16
html5lib==1.1
httplib2==0.22.0
humanize==4.9.0
idna==3.4
imgkit==1.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.4
lunardate==0.2.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.26.3
oauthlib==3.2.2
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
pdfkit==1.0.0
Pillow==10.1.0
pipreqs==0.4.13
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
PyGetWindow==0.0.9
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyluach==2.2.0
PyMeeus==0.5.12
PyMsgBox==1.0.9
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==3.0.1
pyperclip==1.8.2
pyphen==0.14.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
rsa==4.9
s3transfer==0.7.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tornado==6.3.3
trio==0.23.1
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.7
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.12
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
wikipedia==1.4.0
workalendar==17.0.0
wsproto==1.2.0
xhtml2pdf==0.2.13
yarg==0.1.9
zopfli==0.2.3
